Add support for LISP tunneling
LISP is an experimental layer 3 tunneling protocol.  This patch adds
support for LISP tunneling.  Since LISP encapsulated packets do not
carry an Ethernet header, it is removed before encapsulation, and added
with hardcoded source and destination MAC addresses after decapsulation.
The harcoded MAC chosen for this purpose is the locally administered
address 02:00:00:00:00:00.  Flow actions can be used to rewrite this MAC
for correct reception.  As such, this patch is intended to be used for
static network configurations, or with a LISP capable controller.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>